# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  suspended: 'suspended',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  status?: UserStatus;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
  /** @minLength 6 */
  password?: string;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  expired: 'expired',
  suspended: 'suspended',
} as const;

export interface Subscription {
  id: number;
  userId: number;
  status: SubscriptionStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  daysPaid: number;
  remainingTradingDays?: number;
  createdAt: string;
}

export interface PaymentInput {
  phone: string;
  /**
     * @minimum 1
     * @maximum 365
     */
  days: number;
}

export interface PaymentInitResponse {
  checkoutRequestId: string;
  message: string;
  amount?: number;
}

export type MpesaCallbackInputBody = { [key: string]: unknown };

export interface MpesaCallbackInput {
  Body?: MpesaCallbackInputBody;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Payment {
  id: number;
  userId: number;
  phone: string;
  amount: number;
  /** @nullable */
  mpesaReceipt?: string | null;
  status: PaymentStatus;
  days?: number;
  createdAt: string;
}

export interface MasterAccountInput {
  broker: string;
  server: string;
  mt5Login: string;
  investorPassword: string;
}

export type MasterAccountStatus = typeof MasterAccountStatus[keyof typeof MasterAccountStatus];


export const MasterAccountStatus = {
  connecting: 'connecting',
  connected: 'connected',
  disconnected: 'disconnected',
  error: 'error',
} as const;

export interface MasterAccount {
  id: number;
  userId: number;
  /** @nullable */
  metaapiAccountId?: string | null;
  mt5Login: string;
  broker: string;
  server: string;
  status: MasterAccountStatus;
  createdAt: string;
}

export interface SlaveAccountInput {
  broker: string;
  server: string;
  mt5Login: string;
  investorPassword: string;
}

export type SlaveAccountStatus = typeof SlaveAccountStatus[keyof typeof SlaveAccountStatus];


export const SlaveAccountStatus = {
  connecting: 'connecting',
  connected: 'connected',
  disconnected: 'disconnected',
  suspended: 'suspended',
  error: 'error',
} as const;

export interface SlaveAccount {
  id: number;
  userId: number;
  /** @nullable */
  metaapiAccountId?: string | null;
  /** @nullable */
  subscriberId?: string | null;
  mt5Login: string;
  broker: string;
  server: string;
  status: SlaveAccountStatus;
  createdAt: string;
}

export interface StrategyInput {
  /** @minLength 2 */
  strategyName: string;
  masterAccountId: number;
}

export type StrategyStatus = typeof StrategyStatus[keyof typeof StrategyStatus];


export const StrategyStatus = {
  active: 'active',
  paused: 'paused',
  error: 'error',
} as const;

export interface Strategy {
  id: number;
  userId: number;
  /** @nullable */
  copyfactoryStrategyId?: string | null;
  strategyName: string;
  masterAccountId: number;
  status: StrategyStatus;
  createdAt: string;
}

export interface BindingInput {
  strategyId: number;
  slaveAccountId: number;
  /**
     * @minimum 0.1
     * @maximum 10
     */
  riskMultiplier: number;
}

export type BindingStatus = typeof BindingStatus[keyof typeof BindingStatus];


export const BindingStatus = {
  active: 'active',
  suspended: 'suspended',
} as const;

export interface Binding {
  id: number;
  strategyId: number;
  slaveAccountId: number;
  riskMultiplier: number;
  status: BindingStatus;
  createdAt: string;
}

export interface TradeLog {
  id: number;
  strategyId: number;
  /** @nullable */
  slaveAccountId?: number | null;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  subscription: Subscription;
  masterAccounts: number;
  slaveAccounts: number;
  strategies: number;
  activeBindings: number;
  recentPayments: Payment[];
  recentTradeLogs: TradeLog[];
}

export interface AdminStats {
  totalUsers: number;
  activeSubscriptions: number;
  totalRevenue: number;
  activeSlaveAccounts: number;
  activeStrategies: number;
  totalPayments: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: string;
  status: string;
  /** @nullable */
  subscriptionStatus?: string | null;
  /** @nullable */
  remainingDays?: number | null;
  createdAt: string;
}

export interface AdminSettings {
  id: number;
  dailyFee: number;
  minDays: number;
  maxDays: number;
  metaApiToken?: string | null;
  updatedAt?: string;
}

export interface AdminSettingsUpdate {
  /** @minimum 1 */
  dailyFee?: number;
  /** @minimum 1 */
  minDays?: number;
  /** @maximum 365 */
  maxDays?: number;
  metaApiToken?: string | null;
}