# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /master-accounts:
    get:
      operationId: listMasterAccounts
      tags: [masterAccounts]
      summary: List current user's master accounts
      responses:
        "200":
          description: List of master accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MasterAccount"
    post:
      operationId: createMasterAccount
      tags: [masterAccounts]
      summary: Create a new master MT5 account via MetaApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterAccountInput"
      responses:
        "201":
          description: Master account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterAccount"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /master-accounts/{id}:
    get:
      operationId: getMasterAccount
      tags: [masterAccounts]
      summary: Get a master account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Master account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterAccount"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMasterAccount
      tags: [masterAccounts]
      summary: Delete a master account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SLAVE ACCOUNTS
  /slave-accounts:
    get:
      operationId: listSlaveAccounts
      tags: [slaveAccounts]
      summary: List current user's slave accounts
      responses:
        "200":
          description: List of slave accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SlaveAccount"
    post:
      operationId: createSlaveAccount
      tags: [slaveAccounts]
      summary: Create a new slave MT5 account via MetaApi CopyFactory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlaveAccountInput"
      responses:
        "201":
          description: Slave account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaveAccount"
        "400":
          description: Invalid request or subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /slave-accounts/{id}:
    delete:
      operationId: deleteSlaveAccount
      tags: [slaveAccounts]
      summary: Delete a slave account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # STRATEGIES
...
  /admin/users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/activate:
    patch:
      operationId: activateUser
      tags: [admin]
      summary: Activate a suspended user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get platform settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings (daily fee, min/max days)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MasterAccountInput:
      type: object
      required: [broker, server, mt5Login, investorPassword]
      properties:
        broker:
          type: string
        server:
          type: string
        mt5Login:
          type: string
        investorPassword:
          type: string
    MasterAccount:
      type: object
      required: [id, userId, mt5Login, broker, server, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        metaapiAccountId:
          type: ["string", "null"]
        mt5Login:
          type: string
        broker:
          type: string
        server:
          type: string
        status:
          type: string
          enum: [connecting, connected, disconnected, error]
        createdAt:
          type: string
          format: date-time

    # SLAVE ACCOUNT
    SlaveAccountInput:
      type: object
      required: [broker, server, mt5Login, investorPassword]
      properties:
        broker:
          type: string
        server:
          type: string
        mt5Login:
          type: string
        investorPassword:
          type: string
    SlaveAccount:
      type: object
      required: [id, userId, mt5Login, broker, server, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        metaapiAccountId:
          type: ["string", "null"]
        subscriberId:
          type: ["string", "null"]
        mt5Login:
          type: string
        broker:
          type: string
        server:
          type: string
        status:
          type: string
          enum: [connecting, connected, disconnected, suspended, error]
        createdAt:
          type: string
          format: date-time

    # STRATEGY
...
    AdminUser:
      type: object
      required: [id, name, email, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
        subscriptionStatus:
          type: ["string", "null"]
        remainingDays:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    AdminSettings:
      type: object
      required: [id, dailyFee, minDays, maxDays]
      properties:
        id:
          type: integer
        dailyFee:
          type: number
        minDays:
          type: integer
        maxDays:
          type: integer
        metaApiToken:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
    AdminSettingsUpdate:
      type: object
      properties:
        dailyFee:
          type: number
          minimum: 1
        minDays:
          type: integer
          minimum: 1
        maxDays:
          type: integer
          maximum: 365
        metaApiToken:
          type: string
          nullable: true