# lib/api-zod/src/generated/types/adminSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettings {
  id: number;
  dailyFee: number;
  minDays: number;
  maxDays: number;
  metaApiToken?: string | null;
  updatedAt?: Date;
}