# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettingsUpdate {
  /** @minimum 1 */
  dailyFee?: number;
  /** @minimum 1 */
  minDays?: number;
  /** @maximum 365 */
  maxDays?: number;
  metaApiToken?: string | null;
}